+++ src/Client/Module.h
@@ -261,15 +261,29 @@ class InputModule : public Module, public TextInputDelegate
         }
 };
 
-class ColourModule : public Module, public TextInputDelegate
+class ColourDelegate : public geode::ColorPickPopupDelegate
+{
+    public:
+        Module* mod = nullptr;
+
+        virtual void updateColor(cocos2d::ccColor4B const& color);
+};
+
+class ColourModule : public Module
 {
     public:
         static inline ColourModule* selected = nullptr;
 
         ccColor3B colour = ccc3(255, 255, 255);
 
+        ColourDelegate* del = nullptr;
+
+        CCSprite* btnSpr = nullptr;
+
         ColourModule(std::string name, std::string id, ccColor3B def)
         {
+            this->name = name;
+            this->id = id;
             colour = def;
 
             this->load();
@@ -280,9 +294,9 @@ class ColourModule : public Module, public TextInputDelegate
             ImVec2 pos = ImGui::GetCursorPos();
 
             auto v = name;
-            name = "";            
+            //name = "";            
             auto res = Module::Draw(tileSize);
-            name = v;
+            //name = v;
 
             float height = tileSize.y - 10;
 
@@ -301,16 +315,101 @@ class ColourModule : public Module, public TextInputDelegate
             colour = geode::prelude::Mod::get()->getSavedValue<ccColor3B>(id + "_value", colour);
         }
 
+        void onPickColourAndroid(CCObject* sender)
+        {
+            auto v = as<ColourModule*>(as<CCNode*>(sender)->getUserData());
+
+            if (!v->del)
+            {
+                v->del = new ColourDelegate();
+                v->del->mod = v;
+            }
+
+            auto men = geode::ColorPickPopup::create(v->colour);
+
+            men->setDelegate(v->del);
+            men->show();
+        }
+
         void makeAndroid(CCMenu* menu, CCPoint pos)
         {
             auto label = CCLabelBMFont::create(name.c_str(), "bigFont.fnt");
             label->setAnchorPoint(ccp(0, 0.5f));
             label->setScale(0.575f);
-            label->setPosition(pos - ccp(10, 0));
-            label->limitLabelWidth(70, 0.575f, 0.1f);
+            label->setPosition(pos + ccp(-15, 0));
+            label->limitLabelWidth(110, 0.575f, 0.1f);
+
+            btnSpr = CCSprite::createWithSpriteFrameName("GJ_colorBtn_001.png");
+            btnSpr->setColor(colour);
+            btnSpr->setScale(0.75f);
+            auto btn = CCMenuItemSpriteExtra::create(btnSpr, menu, menu_selector(ColourModule::onPickColourAndroid));
+
+            btn->setUserData(this);
+            btn->setPosition(pos + ccp(135, 0) + ccp(-15, 0));
+            menu->addChild(btn);
 
             //menu->addChild(input);
             
+            menu->addChild(label);
+        }
+};
+
+class SliderModule : public Module
+{
+    public:
+        Slider* slider = nullptr;
+        float value = 1.0f;
+
+        SliderModule(std::string name, std::string id, float def)
+        {
+            this->name = name;
+            this->id = id;
+            value = def;
+
+            this->load();
+        }
+
+        void save()
+        {
+            geode::prelude::Mod::get()->setSavedValue<float>(id + "_value", value);
+        }
+
+        void load()
+        {
+            value = geode::prelude::Mod::get()->getSavedValue<float>(id + "_value", value);
+        }
+
+        void onSliderChanged(CCObject* sender)
+        {
+            auto x = as<SliderModule*>(as<SliderThumb*>(sender)->getUserData());
+
+            x->value = as<SliderThumb*>(sender)->getValue();
+            
+            x->save();
+        }
+
+        void makeAndroid(CCMenu* menu, CCPoint pos)
+        {
+            auto label = CCLabelBMFont::create(name.c_str(), "bigFont.fnt");
+            label->setAnchorPoint(ccp(0, 0.5f));
+            label->setScale(0.575f);
+            label->setPosition(pos + ccp(-15, 0));
+            label->limitLabelWidth(80, 0.575f, 0.1f);
+
+            slider = Slider::create(menu, menu_selector(SliderModule::onSliderChanged));
+            slider->getThumb()->setValue(value);
+            slider->updateBar();
+            slider->getThumb()->setUserData(this);
+
+            slider->setAnchorPoint(ccp(0, 0));
+            slider->setScale(0.5f);
+            slider->setScaleX(0.35f);
+            slider->setPosition(pos + ccp(135, 0) + ccp(-15 - 14, 0));
+            
+            as<CCSprite*>(slider->getThumb()->getChildren()->objectAtIndex(0))->setScaleX(0.5f / 0.35f);
+            as<CCSprite*>(slider->getThumb()->getChildren()->objectAtIndex(1))->setScaleX(0.5f / 0.35f);
+            
+            menu->addChild(slider);
             menu->addChild(label);
         }
 };